Move certificate interface into dedicated package to avoid cycle dependency
